+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -507,9 +507,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-16T09:24:13.490Z"
+    created: Optional[str] = "2026-07-16T11:02:25.452Z"
 
-    updated: Optional[str] = "2026-07-16T09:24:13.490Z"
+    updated: Optional[str] = "2026-07-16T11:02:25.452Z"
 
     updated_by_id: OptionalNullable[str] = UNSET