+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXN3VS7Q653EBHABZB76FEBF"
+        "01KXN9FG8C8DTNFNH7XBTP487Q"
     )
     r"""The unique identifier of the data source"""